requirements.txt
Django==3.2.6
django-admin==2.0.1
django-allauth==0.45.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.12.1
gunicorn==20.1.0
httpie==2.4.0
model-bakery==1.3.2
pytest==6.2.4
pytest_check==1.0.2
psycopg2-binary>=2.8
django-extensions==3.1.3
Pillow==8.4.0
apiclient~=1.0.4